refactor: extract cdp-element-resolve.ts from cdp-dispatch.ts
Move cdpResolveAndPrepare (83 LOC, risk 0.68), resolveElement, and
buildCDPResult into dedicated module. These handle page-injected element
resolution, separate from CDP protocol dispatch.
